Index Implementation
* Clean up Events
* Removed `addEventListener`
* Added Observers
* Notification handlers
